+++ python/pathway/xpacks/llm/llms.py
@@ -12,7 +12,6 @@ import json
 import logging
 import uuid
 from abc import abstractmethod
-from collections.abc import Coroutine
 from typing import Any, Iterable, Literal
 
 import pathway as pw
@@ -55,11 +54,6 @@ class BaseChat(pw.UDF):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.kwargs = {}
-
-    @abstractmethod
-    def __wrapped__(
-        self, messages: list[dict] | pw.Json, *args, **kwargs
-    ) -> str | Coroutine[Any, Any, str | None] | Any: ...
 
     @abstractmethod
     def _accepts_call_arg(self, arg_name: str) -> bool: ...
@@ -462,6 +456,8 @@ class HFPipelineChat(BaseChat):
             `the HuggingFace documentation
             <https://huggingface.co/docs/transformers/en/main_classes/pipelines#transformers.TextGenerationPipeline.__call__>`_.
         device: defines which device will be used to run the Pipeline
+        batch_size: maximum size of a single batch to be sent to the chat. Bigger
+            batches may reduce the time needed for generating the answer, especially on GPU.
         pipeline_kwargs: kwargs accepted during initialization of HuggingFace Pipeline.
             For possible arguments check
             `the HuggingFace documentation <https://huggingface.co/docs/transformers/en/main_classes/pipelines#transformers.pipeline>`_.
@@ -488,32 +484,92 @@ class HFPipelineChat(BaseChat):
         model: str | None = "gpt2",
         call_kwargs: dict = {},
         device: str = "cpu",
+        batch_size: int = 32,
         **pipeline_kwargs,
     ):
         with optional_imports("xpack-llm-local"):
             import transformers
 
-        super().__init__()
+        super().__init__(max_batch_size=batch_size)
         self.pipeline = transformers.pipeline(
             model=model, device=device, **pipeline_kwargs
         )
         self.tokenizer = self.pipeline.tokenizer
-        self.kwargs.update(call_kwargs)
+        self.kwargs = {"batch_size": batch_size, **call_kwargs}
 
-    def __wrapped__(self, messages: list[dict] | pw.Json | str, **kwargs) -> str | None:
-        if isinstance(messages, str):
-            messages_decoded: list[dict] | str = messages
-        else:
-            messages_decoded = _prepare_messages(messages)
+    def __wrapped__(
+        self, messages_list: list[list[dict] | pw.Json | str], **kwargs
+    ) -> list[str | None]:
 
-        kwargs = {**self.kwargs, **kwargs}
+        def decode_messages(messages: list[dict] | pw.Json | str) -> list[dict] | str:
+            if isinstance(messages, str):
+                return messages
+            else:
+                return _prepare_messages(messages)
+
+        messages_decoded_list = [
+            decode_messages(messages) for messages in messages_list
+        ]
         kwargs = _extract_value_inside_dict(kwargs)
+        constant_kwargs = {}
+        per_row_kwargs = {}
 
-        output = self.pipeline(messages_decoded, **kwargs)
-        result = output[0]["generated_text"]
-        if isinstance(result, list):
-            result = result[-1]["content"]
-        return result
+        if kwargs:
+            for key, values in kwargs.items():
+                v = values[0]
+                if all(value == v for value in values):
+                    constant_kwargs[key] = v
+                else:
+                    per_row_kwargs[key] = values
+
+        def decode_output(output) -> str | None:
+            result = output[0]["generated_text"]
+            if isinstance(result, list):
+                return result[-1]["content"]
+            else:
+                return result
+
+        # if kwargs are not the same for every message we cannot batch them
+        # huggingface does not allow batching if tokenizer or tokenizer.pad_token_id is None
+        if (
+            per_row_kwargs
+            or self.pipeline.tokenizer is None
+            or self.pipeline.tokenizer.pad_token_id is None
+        ):
+
+            def infer_single(messages, kwargs) -> str | None:
+                kwargs = {**self.kwargs, **constant_kwargs, **kwargs}
+                output = self.pipeline(messages, **kwargs)
+                return decode_output(output)
+
+            list_of_per_row_kwargs = [
+                dict(zip(per_row_kwargs, values))
+                for values in zip(*per_row_kwargs.values())
+            ]
+
+            if list_of_per_row_kwargs:
+                result_list = [
+                    infer_single(messages, kwargs)
+                    for messages, kwargs in zip(
+                        messages_decoded_list, list_of_per_row_kwargs
+                    )
+                ]
+            else:
+                result_list = [
+                    infer_single(messages, {}) for messages in messages_decoded_list
+                ]
+
+            return result_list
+
+        else:
+            kwargs = {**self.kwargs, **constant_kwargs}
+            output_list = self.pipeline(messages_decoded_list, **kwargs)
+
+            if output_list is None:
+                return [None] * len(messages_list)
+
+            result_list = [decode_output(output) for output in output_list]
+            return result_list
 
     def crop_to_max_length(
         self, input_string: pw.ColumnExpression, max_prompt_length: int = 500